Fix: Page 404 only shows in language deafult
