+++ pages/404.tsx
@@ -1,9 +1,23 @@
-// to detect language and automatically redirect to the approprate/[locale] page
-// import { Redirect } from '../lib/redirect'
-// export default Redirect
+import { useRouter } from "next/router"
+import Error404 from "../app/components/error404/Error404"
+import common from "../public/locales/en/common.json"
 
-// to keep this root page with the defaultLocale
-import NotFound, { getStaticProps } from "./[locale]/404"
+const NotFound = () => {
+  const router = useRouter()
+  const translation = common
+  return (
+    <main style={{ display: "flex" }}>
+      <Error404
+        text={translation["message-error-404"]}
+        buttonTextEnglish={translation["name-button-404-en"]}
+        buttonTextSpanish={translation["name-button-404-es"]}
+        styleButton="cta"
+        scrollToSection=""
+        handleClickEnglish={() => router.replace("/")}
+        handleClickSpanish={() => router.replace("/es")}
+      />
+    </main>
+  )
+}
 
 export default NotFound
-export { getStaticProps }
