+++ pages/_error.tsx
@@ -1,6 +1,6 @@
 import { NextPage } from "next"
 import { ErrorPageProps } from "../app/constants/types/components_props/types"
-import NotFound from "./[locale]/404"
+import NotFound from "./404"
 
 const ErrorPage: NextPage<ErrorPageProps> = ({ statusCode }) => {
   if (statusCode === 404) {
