+++ tsconfig.json
@@ -28,7 +28,7 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
-    "pages/[locale]/404.tsx",
+    "pages/404.tsx",
     ".next/types/**/*.ts"
   ],
   "exclude": [
